+++ spec/presenters/admin_artist_list_spec.rb
@@ -1,7 +1,7 @@
 require 'rails_helper'
 require 'capybara/rspec'
 
-describe AdminArtistList, elasticsearch: false do
+describe AdminArtistList do
   include PresenterSpecHelpers
 
   let(:active_artists) do
